Feat: Changed output for empty search. Made modeify call tableify dynamically. Refactor: Organized files into folders. Created helper functions for handlers
