+++ mock/src/components/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import '../styles/App.css';
-import { LoginButton } from './LoginButton';
-import REPL from './REPL';
+import { LoginButton } from './Elements/LoginButton';
+import REPL from './REPL/REPL';
 
 /**
  * This is the highest level component!
