+++ mock/src/components/REPL/REPL.tsx
@@ -1,0 +1,16 @@
+import { useState } from "react";
+import "../../styles/main.css";
+import { REPLHistory } from "./REPLHistory";
+import { REPLInput } from "./REPLInput";
+
+export default function REPL() {
+  const [history, setHistory] = useState<JSX.Element[]>([]);
+
+  return (
+    <div className="repl">
+      <REPLHistory history={history}/>
+      <hr></hr>
+      <REPLInput history={history} setHistory={setHistory} />
+    </div>
+  );
+}
