+++ mock/src/components/REPL/REPLHistory.tsx
@@ -1,0 +1,16 @@
+import "../../styles/main.css";
+
+interface REPLHistoryProps {
+  history: JSX.Element[];
+}
+export function REPLHistory(props: REPLHistoryProps) {
+  
+
+  return (
+    <div className="repl-history">
+      {props.history.map((command, index) => (
+        command
+      ))}
+    </div>
+  );
+}
