+++ mock/src/components/REPL/REPLInput.tsx
@@ -1,13 +1,14 @@
 import { Dispatch, SetStateAction, useState } from "react";
-import "../styles/main.css";
-import { ControlledInput } from "./ControlledInput";
+import "../../styles/main.css";
+import { ControlledInput } from "../Elements/ControlledInput";
 import {
   handleMode,
   handleView,
   handleSearch,
   handleLoad,
   REPLFunction,
-} from "./Handlers";
+  REPLFuctionStatefulInputProps,
+} from "../Utilities/Handlers";
 
 interface REPLInputProps {
   history: JSX.Element[];
@@ -16,9 +17,9 @@ interface REPLInputProps {
 
 export function REPLInput(props: REPLInputProps) {
   const [commandString, setCommandString] = useState<string>("");
-  const [modeBrief, setMode] = useState<Boolean>(true);
+  const [brief, setBrief] = useState<Boolean>(true);
   const [count, setCount] = useState<number>(0);
-  const [file, setFile] = useState<String>("");
+  const [file, setFileName] = useState<String>("");
   const [headers, setHeaders] = useState<Boolean>(false);
 
   const commandMap: Map<String, REPLFunction> = new Map<String, REPLFunction>([
@@ -37,15 +38,15 @@ export function REPLInput(props: REPLInputProps) {
     if (commandArgs.length > 0) {
       const command = commandMap.get(commandArgs[0]);
       if (command != undefined) {
-        commandReturn = command(
-          commandArgs,
-          modeBrief,
-          setMode,
-          file,
-          setFile,
-          headers,
-          setHeaders
-        );
+        const statefuls: REPLFuctionStatefulInputProps = {
+          brief: brief,
+          setBrief: setBrief,
+          file: file,
+          setFileName: setFileName,
+          headers: headers,
+          setHeaders: setHeaders,
+        };
+        commandReturn = command(commandArgs, statefuls);
       } else {
         commandReturn = <p>{"Command does not exist :("}</p>;
       }
@@ -57,16 +58,8 @@ export function REPLInput(props: REPLInputProps) {
     setCommandString("");
   }
 
-  /**
-   * We suggest breaking down this component into smaller components, think about the individual pieces
-   * of the REPL and how they connect to each other...
-   */
   return (
     <div className="repl-input">
-      {/* This is a comment within the JSX. Notice that it's a TypeScript comment wrapped in
-            braces, so that React knows it should be interpreted as TypeScript */}
-      {/* I opted to use this HTML tag; you don't need to. It structures multiple input fields
-            into a single unit, which makes it easier for screenreaders to navigate. */}
       <fieldset>
         <legend>Enter a command:</legend>
         <ControlledInput
